Renames mobile-nav component
